Modify blog-list ssr
In order to work with either remote OR local content, this needs to
generate all data at this stage, before passing it to the template. This
commit updates the blog-list render function to do this, but notably
does NOT update the template part (that would come in a future commit).