Fix language tax update when saving content in the block editor
When creating content locally with the latest version of shiro theme,
any attempts to update newly-created posts break with the following
error:

    Fatal error: Uncaught Error: Call to undefined function wp_create_term()
    in /srv/www/wikimediafoundation/public_html/wp-content/themes/shiro/inc/taxonomy-content-language.php on line 80

This is happening because in the REST API endpoint hit to update content
in the block editor, wp-admin/includes/taxonomy.php isn't loaded by
default. This change ensures that this file is required whenever trying
to update language taxonomy terms.