Prevent error accessing unregistered meta value
The "wmf_hide_title" value wasn't registered for any post types other
than Pages, so trying to access it with the `useMeta` hook was throwing
errors for any users who weren't super-admins. This moves the hook
inside a component which is conditionally rendered, so that it isn't
being accessed where it shouldn't be.