88: Fix JS error blocking execution of unrelated plugin code.
The filters we introduced in #55 work well, but the code that manages them
still runs on pages outside the News section. Because the filter interface
doesn't exist on other pages, these `querySelector` calls returned `null`,
and attempts to manipulate those null values fail.

This PR slightly refactors the existing code to add guards around the node
queries so that we only run code when the corresponding HTML's on the page.

Fixes #88.