920: Exit immediately if theme file is directly accessed
VIP runtime logs reveal that some requests are made directly to theme
files, which is possibly benign but also possibly as a penetration testing
tactic. The file will not load if WordPress is not running, because wp_head
will not be defined, but this guard code ensures that the page exits
immediately if any top-level theme file is accessed.