1012: Improve build performance by avoiding SVG processing
This commit excludes url() css-loader parsing by adding the # character
so that the loader will ignore the entire file call; otherwise it would
try to load and process the import, which is slow. The stylesheets today
rely on the files being where we expect them to be on-disk, so leavingt
the processing as-is also avoids accidentally changing behavior.

We then use "sed" to rewrite the file contents after render.