1012: Do not process url() as import in legacy stylesheets
It is prohibitively slow for Webpack to import and parse all of the
files referenced in the existing stylesheets, and the CSS is written
to assume the files exist properly relative to the built asset -- we
do not need to process anything ourselves. This url filter function
lets us opt specific directories out of css-loader processing.