Guard against unset Slide.slide variable in homepage carousel
While validating the deploy on August 7 we briefly observed a JS error
that `{ slide }` could not be destructured from a null reference. While
we suspect this was a transient situation due to the rollout of the
updated page content and markup, we _could_ add an existence check here
prior to destructuring to be more confident the object exists.